// src/api/generated/auntyKoSchoolAPI.schemas.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Aunty ko SchoolAPI
 * Lu Srajay banaidiye tero SWAGGER
 * OpenAPI spec version: 1.0
 */
export interface Form {
  leadId?: number;
  leadName?: string;
  leadEmail?: string;
  leadPhone?: string;
  leadEnquiry?: string;
}

export interface FormUpdate {
  leadName?: string;
  leadEmail?: string;
  leadPhone?: string;
}

export interface User {
  userId?: number;
  userName?: string;
  role?: string;
}

export interface UserCreate {
  userName: string;
  password: string;
  cpassword?: string;
  role?: string;
}

export interface UserUpdate {
  userName?: string;
  password?: string;
  role?: string;
}

export interface UserLogin {
  userName: string;
  password: string;
}